+++ backend/src/main/java/com/back/domain/product/product/controller/ApiV1ProductController.java
@@ -3,17 +3,14 @@ package com.back.domain.product.product.controller;
 import com.back.domain.product.product.dto.ProductDto;
 import com.back.domain.product.product.entity.Product;
 import com.back.domain.product.product.service.ProductService;
-import com.back.global.rsData.RsData;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
@@ -23,37 +20,6 @@ import java.util.List;
 @Tag(name = "ApiV1ProductController", description = "API 상품 컨트롤러")
 public class ApiV1ProductController {
     private final ProductService productService;
-
-    record ProductCreateReqBody(
-            @NotBlank
-            @Size(min = 2, max = 100)
-            String name,
-            @Min(100)
-            @Max(1000000)
-            int price,
-            @NotBlank
-            @Size(min = 2, max = 500)
-            String description,
-            @Min(1)
-            @Max(10000)
-            int stock
-    ) {}
-
-    @PostMapping
-    @Transactional
-    @Operation(summary = "생성")
-    public RsData<ProductDto> create(
-            @Valid @RequestBody ProductCreateReqBody reqBody
-    ) {
-        Product product = productService.create(reqBody.name, reqBody.price, reqBody.description, reqBody.stock);
-
-        return new RsData<>(
-                "201-1",
-                "%d번 상품이 등록되었습니다.".formatted(product.getId()),
-                new ProductDto(product)
-        );
-    }
-
 
     @GetMapping
     @Transactional(readOnly = true)
@@ -72,52 +38,5 @@ public class ApiV1ProductController {
     public ProductDto getItem(@PathVariable int id) {
         Product product = productService.findById(id).get();
         return new ProductDto(product);
-    }
-
-
-    @DeleteMapping("/{id}")
-    @Transactional
-    @Operation(summary = "삭제")
-    public RsData<Void> delete(@PathVariable int id) {
-        Product product = productService.findById(id).get();
-        productService.delete(product);
-
-        return new RsData<>(
-                "200-1",
-                "%d번 상품이 삭제되었습니다.".formatted(id)
-        );
-    }
-
-
-    record ProductUpdateReqBody(
-            @NotBlank
-            @Size(min = 2, max = 100)
-            String name,
-            @Min(100)
-            @Max(1000000)
-            int price,
-            @NotBlank
-            @Size(min = 2, max = 500)
-            String description,
-            @Min(1)
-            @Max(10000)
-            int stock
-    ) {}
-
-
-    @PutMapping("/{id}")
-    @Transactional
-    @Operation(summary = "수정")
-    public RsData<Void> update(
-            @PathVariable int id,
-            @Valid @RequestBody ProductUpdateReqBody reqBody
-    ) {
-        Product product = productService.findById(id).get();
-        productService.modify(product, reqBody.name, reqBody.price, reqBody.description, reqBody.stock);
-
-        return new RsData<>(
-                "200-1",
-                "%d번 상품이 수정되었습니다.".formatted(id)
-        );
     }
 }
